Fix: Align `kinds-c` cli arg handling with ctags
Should follow the behaviour described in the ctags documentation
```
Each letter or group of letters may be preceded by either ‘+’ to add it to, or ‘-’ to remove it from, the default set. In the absence of any preceding ‘+’ or ‘-’ sign, only those kinds explicitly listed in kinds will be included in the output (i.e. overriding the default for the specified language).
```

`from_string_legacy` does not confirm to ctags docuemtation and needs to
be removed altogether in future.

Also fixes deviation from ctags about the options accepted by `kinds-c`