Update default value of `recurse` cli option to `true`
The default value was changed from `true` to `false` in a recent commit
to bring defaults closer to those of ctags. But its proving to be
difficult to match ctags behaviour in all cases. For eg to match the
behaviour of `ctags *`. ctags sets `recurse` to true for this case
somehow but not when user supplies a mix of files and directories
themselves instead of via glob expansion by the shell.
Defaulting `recurse` to `true` seems to be a better default than the
other way around if we cannot match ctags behaviour in all cases